+++ CHANGELOG.md
@@ -5,7 +5,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## \[Unreleased\]
 
 ### Added
+<<<<<<< HEAD
 -  A new configuration field `targetArcFactor` has been added to `NetworkConfig`, allowing test scenarios to be written with 0-arc conductors ([#300](https://github.com/holochain/tryorama/pull/300))
+=======
+- A new method `storageArc` that polls for network metrics and to check if an agent's storage arc is equal to a desired storage arc for a given dna hash. ([#300](https://github.com/holochain/tryorama/pull/301))
+>>>>>>> 4812263 (chore: changelog)
 
 ### Removed
 ### Changed